Clarify that api routes should be inside the pages folder
It's quite clear that api routes should be inside the pages folder, however, if you quickly glance over the second sentence you could easily miss it like I did. 

Also use `./` to refer to the root folder as that's how it's done elsewhere in the readme e.g. in the setup section.